List available Arrow C++ compute functions

This function lists the names of all available Arrow C++ library compute functions. These can be called by passing to call_function(), or they can be called by name with an arrow_ prefix inside a dplyr verb.

list_compute_functions(pattern = NULL, ...)

Arguments

pattern
Optional regular expression to filter the function list

...
Additional parameters passed to grep()

Value

A character vector of available Arrow C++ function names

Details

The resulting list describes the capabilities of your arrow build. Some functions, such as string and regular expression functions, require optional build-time C++ dependencies. If your arrow package was not compiled with those features enabled, those functions will not appear in this list.

Some functions take options that need to be passed when calling them (in a list called options). These options require custom handling in C++; many functions already have that handling set up but not all do. If you encounter one that needs special handling for options, please report an issue.

Note that this list does not enumerate all of the R bindings for these functions. The package includes Arrow methods for many base R functions that can be called directly on Arrow objects, as well as some tidyverse-flavored versions available inside dplyr verbs.

Examples

list_compute_functions() 
#>   [1] "abs"                      "abs_checked"             
#>   [3] "acos"                     "acos_checked"            
#>   [5] "add"                      "add_checked"             
#>   [7] "all"                      "and"                     
#>   [9] "and_kleene"               "and_not"                 
#>  [11] "and_not_kleene"           "any"                     
#>  [13] "array_filter"             "array_sort_indices"      
#>  [15] "array_take"               "ascii_center"            
#>  [17] "ascii_is_alnum"           "ascii_is_alpha"          
#>  [19] "ascii_is_decimal"         "ascii_is_lower"          
#>  [21] "ascii_is_printable"       "ascii_is_space"          
#>  [23] "ascii_is_title"           "ascii_is_upper"          
#>  [25] "ascii_lower"              "ascii_lpad"              
#>  [27] "ascii_ltrim"              "ascii_ltrim_whitespace"  
#>  [29] "ascii_reverse"            "ascii_rpad"              
#>  [31] "ascii_rtrim"              "ascii_rtrim_whitespace"  
#>  [33] "ascii_split_whitespace"   "ascii_trim"              
#>  [35] "ascii_trim_whitespace"    "ascii_upper"             
#>  [37] "asin"                     "asin_checked"            
#>  [39] "atan"                     "atan2"                   
#>  [41] "binary_join"              "binary_join_element_wise"
#>  [43] "binary_length"            "binary_replace_slice"    
#>  [45] "bit_wise_and"             "bit_wise_not"            
#>  [47] "bit_wise_or"              "bit_wise_xor"            
#>  [49] "cast"                     "cos"                     
#>  [51] "cos_checked"              "count"                   
#>  [53] "count_substring"          "count_substring_regex"   
#>  [55] "day"                      "day_of_week"             
#>  [57] "day_of_year"              "dictionary_encode"       
#>  [59] "divide"                   "divide_checked"          
#>  [61] "ends_with"                "equal"                   
#>  [63] "extract_regex"            "fill_null"               
#>  [65] "filter"                   "find_substring"          
#>  [67] "find_substring_regex"     "greater"                 
#>  [69] "greater_equal"            "hash_count"              
#>  [71] "hash_min_max"             "hash_sum"                
#>  [73] "hour"                     "if_else"                 
#>  [75] "index"                    "index_in"                
#>  [77] "index_in_meta_binary"     "invert"                  
#>  [79] "is_finite"                "is_in"                   
#>  [81] "is_in_meta_binary"        "is_inf"                  
#>  [83] "is_nan"                   "is_null"                 
#>  [85] "is_valid"                 "iso_calendar"            
#>  [87] "iso_week"                 "iso_year"                
#>  [89] "less"                     "less_equal"              
#>  [91] "list_flatten"             "list_parent_indices"     
#>  [93] "list_value_length"        "match_like"              
#>  [95] "match_substring"          "match_substring_regex"   
#>  [97] "max_element_wise"         "mean"                    
#>  [99] "microsecond"              "millisecond"             
#> [101] "min_element_wise"         "min_max"                 
#> [103] "minute"                   "mode"                    
#> [105] "month"                    "multiply"                
#> [107] "multiply_checked"         "nanosecond"              
#> [109] "negate"                   "negate_checked"          
#> [111] "not_equal"                "or"                      
#> [113] "or_kleene"                "partition_nth_indices"   
#> [115] "power"                    "power_checked"           
#> [117] "project"                  "quantile"                
#> [119] "quarter"                  "replace_substring"       
#> [121] "replace_substring_regex"  "second"                  
#> [123] "shift_left"               "shift_left_checked"      
#> [125] "shift_right"              "shift_right_checked"     
#> [127] "sin"                      "sin_checked"             
#> [129] "sort_indices"             "split_pattern"           
#> [131] "split_pattern_regex"      "starts_with"             
#> [133] "stddev"                   "string_is_ascii"         
#> [135] "strptime"                 "subsecond"               
#> [137] "subtract"                 "subtract_checked"        
#> [139] "sum"                      "take"                    
#> [141] "tan"                      "tan_checked"             
#> [143] "tdigest"                  "unique"                  
#> [145] "utf8_center"              "utf8_is_alnum"           
#> [147] "utf8_is_alpha"            "utf8_is_decimal"         
#> [149] "utf8_is_digit"            "utf8_is_lower"           
#> [151] "utf8_is_numeric"          "utf8_is_printable"       
#> [153] "utf8_is_space"            "utf8_is_title"           
#> [155] "utf8_is_upper"            "utf8_length"             
#> [157] "utf8_lower"               "utf8_lpad"               
#> [159] "utf8_ltrim"               "utf8_ltrim_whitespace"   
#> [161] "utf8_replace_slice"       "utf8_reverse"            
#> [163] "utf8_rpad"                "utf8_rtrim"              
#> [165] "utf8_rtrim_whitespace"    "utf8_slice_codeunits"    
#> [167] "utf8_split_whitespace"    "utf8_trim"               
#> [169] "utf8_trim_whitespace"     "utf8_upper"              
#> [171] "value_counts"             "variance"                
#> [173] "xor"                      "year"                    
list_compute_functions(pattern = "^UTF8", ignore.case = TRUE)
#>  [1] "utf8_center"           "utf8_is_alnum"         "utf8_is_alpha"        
#>  [4] "utf8_is_decimal"       "utf8_is_digit"         "utf8_is_lower"        
#>  [7] "utf8_is_numeric"       "utf8_is_printable"     "utf8_is_space"        
#> [10] "utf8_is_title"         "utf8_is_upper"         "utf8_length"          
#> [13] "utf8_lower"            "utf8_lpad"             "utf8_ltrim"           
#> [16] "utf8_ltrim_whitespace" "utf8_replace_slice"    "utf8_reverse"         
#> [19] "utf8_rpad"             "utf8_rtrim"            "utf8_rtrim_whitespace"
#> [22] "utf8_slice_codeunits"  "utf8_split_whitespace" "utf8_trim"            
#> [25] "utf8_trim_whitespace"  "utf8_upper"           
list_compute_functions(pattern = "^is", invert = TRUE)
#>   [1] "abs"                      "abs_checked"             
#>   [3] "acos"                     "acos_checked"            
#>   [5] "add"                      "add_checked"             
#>   [7] "all"                      "and"                     
#>   [9] "and_kleene"               "and_not"                 
#>  [11] "and_not_kleene"           "any"                     
#>  [13] "array_filter"             "array_sort_indices"      
#>  [15] "array_take"               "ascii_center"            
#>  [17] "ascii_is_alnum"           "ascii_is_alpha"          
#>  [19] "ascii_is_decimal"         "ascii_is_lower"          
#>  [21] "ascii_is_printable"       "ascii_is_space"          
#>  [23] "ascii_is_title"           "ascii_is_upper"          
#>  [25] "ascii_lower"              "ascii_lpad"              
#>  [27] "ascii_ltrim"              "ascii_ltrim_whitespace"  
#>  [29] "ascii_reverse"            "ascii_rpad"              
#>  [31] "ascii_rtrim"              "ascii_rtrim_whitespace"  
#>  [33] "ascii_split_whitespace"   "ascii_trim"              
#>  [35] "ascii_trim_whitespace"    "ascii_upper"             
#>  [37] "asin"                     "asin_checked"            
#>  [39] "atan"                     "atan2"                   
#>  [41] "binary_join"              "binary_join_element_wise"
#>  [43] "binary_length"            "binary_replace_slice"    
#>  [45] "bit_wise_and"             "bit_wise_not"            
#>  [47] "bit_wise_or"              "bit_wise_xor"            
#>  [49] "cast"                     "cos"                     
#>  [51] "cos_checked"              "count"                   
#>  [53] "count_substring"          "count_substring_regex"   
#>  [55] "day"                      "day_of_week"             
#>  [57] "day_of_year"              "dictionary_encode"       
#>  [59] "divide"                   "divide_checked"          
#>  [61] "ends_with"                "equal"                   
#>  [63] "extract_regex"            "fill_null"               
#>  [65] "filter"                   "find_substring"          
#>  [67] "find_substring_regex"     "greater"                 
#>  [69] "greater_equal"            "hash_count"              
#>  [71] "hash_min_max"             "hash_sum"                
#>  [73] "hour"                     "if_else"                 
#>  [75] "index"                    "index_in"                
#>  [77] "index_in_meta_binary"     "invert"                  
#>  [79] "less"                     "less_equal"              
#>  [81] "list_flatten"             "list_parent_indices"     
#>  [83] "list_value_length"        "match_like"              
#>  [85] "match_substring"          "match_substring_regex"   
#>  [87] "max_element_wise"         "mean"                    
#>  [89] "microsecond"              "millisecond"             
#>  [91] "min_element_wise"         "min_max"                 
#>  [93] "minute"                   "mode"                    
#>  [95] "month"                    "multiply"                
#>  [97] "multiply_checked"         "nanosecond"              
#>  [99] "negate"                   "negate_checked"          
#> [101] "not_equal"                "or"                      
#> [103] "or_kleene"                "partition_nth_indices"   
#> [105] "power"                    "power_checked"           
#> [107] "project"                  "quantile"                
#> [109] "quarter"                  "replace_substring"       
#> [111] "replace_substring_regex"  "second"                  
#> [113] "shift_left"               "shift_left_checked"      
#> [115] "shift_right"              "shift_right_checked"     
#> [117] "sin"                      "sin_checked"             
#> [119] "sort_indices"             "split_pattern"           
#> [121] "split_pattern_regex"      "starts_with"             
#> [123] "stddev"                   "string_is_ascii"         
#> [125] "strptime"                 "subsecond"               
#> [127] "subtract"                 "subtract_checked"        
#> [129] "sum"                      "take"                    
#> [131] "tan"                      "tan_checked"             
#> [133] "tdigest"                  "unique"                  
#> [135] "utf8_center"              "utf8_is_alnum"           
#> [137] "utf8_is_alpha"            "utf8_is_decimal"         
#> [139] "utf8_is_digit"            "utf8_is_lower"           
#> [141] "utf8_is_numeric"          "utf8_is_printable"       
#> [143] "utf8_is_space"            "utf8_is_title"           
#> [145] "utf8_is_upper"            "utf8_length"             
#> [147] "utf8_lower"               "utf8_lpad"               
#> [149] "utf8_ltrim"               "utf8_ltrim_whitespace"   
#> [151] "utf8_replace_slice"       "utf8_reverse"            
#> [153] "utf8_rpad"                "utf8_rtrim"              
#> [155] "utf8_rtrim_whitespace"    "utf8_slice_codeunits"    
#> [157] "utf8_split_whitespace"    "utf8_trim"               
#> [159] "utf8_trim_whitespace"     "utf8_upper"              
#> [161] "value_counts"             "variance"                
#> [163] "xor"                      "year"