List available Arrow C++ compute functions

This function lists the names of all available Arrow C++ library compute functions. These can be called by passing to call_function(), or they can be called by name with an arrow_ prefix inside a dplyr verb.

list_compute_functions(pattern = NULL, ...)

Arguments

pattern
Optional regular expression to filter the function list

...
Additional parameters passed to grep()

Value

A character vector of available Arrow C++ function names

Details

The resulting list describes the capabilities of your arrow build. Some functions, such as string and regular expression functions, require optional build-time C++ dependencies. If your arrow package was not compiled with those features enabled, those functions will not appear in this list.

Some functions take options that need to be passed when calling them (in a list called options). These options require custom handling in C++; many functions already have that handling set up but not all do. If you encounter one that needs special handling for options, please report an issue.

Note that this list does not enumerate all of the R bindings for these functions. The package includes Arrow methods for many base R functions that can be called directly on Arrow objects, as well as some tidyverse-flavored versions available inside dplyr verbs.

Examples

list_compute_functions() 
#>   [1] "abs"                     "abs_checked"            
#>   [3] "add"                     "add_checked"            
#>   [5] "all"                     "and"                    
#>   [7] "and_kleene"              "and_not"                
#>   [9] "and_not_kleene"          "any"                    
#>  [11] "array_filter"            "array_sort_indices"     
#>  [13] "array_take"              "ascii_is_alnum"         
#>  [15] "ascii_is_alpha"          "ascii_is_decimal"       
#>  [17] "ascii_is_lower"          "ascii_is_printable"     
#>  [19] "ascii_is_space"          "ascii_is_title"         
#>  [21] "ascii_is_upper"          "ascii_lower"            
#>  [23] "ascii_ltrim"             "ascii_ltrim_whitespace" 
#>  [25] "ascii_reverse"           "ascii_rtrim"            
#>  [27] "ascii_rtrim_whitespace"  "ascii_split_whitespace" 
#>  [29] "ascii_trim"              "ascii_trim_whitespace"  
#>  [31] "ascii_upper"             "binary_length"          
#>  [33] "cast"                    "count"                  
#>  [35] "dictionary_encode"       "divide"                 
#>  [37] "divide_checked"          "equal"                  
#>  [39] "extract_regex"           "fill_null"              
#>  [41] "filter"                  "find_substring"         
#>  [43] "greater"                 "greater_equal"          
#>  [45] "hash_count"              "hash_min_max"           
#>  [47] "hash_sum"                "index"                  
#>  [49] "index_in"                "index_in_meta_binary"   
#>  [51] "invert"                  "is_finite"              
#>  [53] "is_in"                   "is_in_meta_binary"      
#>  [55] "is_inf"                  "is_nan"                 
#>  [57] "is_null"                 "is_valid"               
#>  [59] "less"                    "less_equal"             
#>  [61] "list_flatten"            "list_parent_indices"    
#>  [63] "list_value_length"       "match_like"             
#>  [65] "match_substring"         "match_substring_regex"  
#>  [67] "mean"                    "min_max"                
#>  [69] "mode"                    "multiply"               
#>  [71] "multiply_checked"        "negate"                 
#>  [73] "negate_checked"          "not_equal"              
#>  [75] "or"                      "or_kleene"              
#>  [77] "partition_nth_indices"   "power"                  
#>  [79] "power_checked"           "project"                
#>  [81] "quantile"                "replace_substring"      
#>  [83] "replace_substring_regex" "sort_indices"           
#>  [85] "split_pattern"           "split_pattern_regex"    
#>  [87] "stddev"                  "string_is_ascii"        
#>  [89] "strptime"                "subtract"               
#>  [91] "subtract_checked"        "sum"                    
#>  [93] "take"                    "tdigest"                
#>  [95] "unique"                  "utf8_is_alnum"          
#>  [97] "utf8_is_alpha"           "utf8_is_decimal"        
#>  [99] "utf8_is_digit"           "utf8_is_lower"          
#> [101] "utf8_is_numeric"         "utf8_is_printable"      
#> [103] "utf8_is_space"           "utf8_is_title"          
#> [105] "utf8_is_upper"           "utf8_length"            
#> [107] "utf8_lower"              "utf8_ltrim"             
#> [109] "utf8_ltrim_whitespace"   "utf8_reverse"           
#> [111] "utf8_rtrim"              "utf8_rtrim_whitespace"  
#> [113] "utf8_split_whitespace"   "utf8_trim"              
#> [115] "utf8_trim_whitespace"    "utf8_upper"             
#> [117] "value_counts"            "variance"               
#> [119] "xor"                    
list_compute_functions(pattern = "^UTF8", ignore.case = TRUE)
#>  [1] "utf8_is_alnum"         "utf8_is_alpha"         "utf8_is_decimal"      
#>  [4] "utf8_is_digit"         "utf8_is_lower"         "utf8_is_numeric"      
#>  [7] "utf8_is_printable"     "utf8_is_space"         "utf8_is_title"        
#> [10] "utf8_is_upper"         "utf8_length"           "utf8_lower"           
#> [13] "utf8_ltrim"            "utf8_ltrim_whitespace" "utf8_reverse"         
#> [16] "utf8_rtrim"            "utf8_rtrim_whitespace" "utf8_split_whitespace"
#> [19] "utf8_trim"             "utf8_trim_whitespace"  "utf8_upper"           
list_compute_functions(pattern = "^is", invert = TRUE)
#>   [1] "abs"                     "abs_checked"            
#>   [3] "add"                     "add_checked"            
#>   [5] "all"                     "and"                    
#>   [7] "and_kleene"              "and_not"                
#>   [9] "and_not_kleene"          "any"                    
#>  [11] "array_filter"            "array_sort_indices"     
#>  [13] "array_take"              "ascii_is_alnum"         
#>  [15] "ascii_is_alpha"          "ascii_is_decimal"       
#>  [17] "ascii_is_lower"          "ascii_is_printable"     
#>  [19] "ascii_is_space"          "ascii_is_title"         
#>  [21] "ascii_is_upper"          "ascii_lower"            
#>  [23] "ascii_ltrim"             "ascii_ltrim_whitespace" 
#>  [25] "ascii_reverse"           "ascii_rtrim"            
#>  [27] "ascii_rtrim_whitespace"  "ascii_split_whitespace" 
#>  [29] "ascii_trim"              "ascii_trim_whitespace"  
#>  [31] "ascii_upper"             "binary_length"          
#>  [33] "cast"                    "count"                  
#>  [35] "dictionary_encode"       "divide"                 
#>  [37] "divide_checked"          "equal"                  
#>  [39] "extract_regex"           "fill_null"              
#>  [41] "filter"                  "find_substring"         
#>  [43] "greater"                 "greater_equal"          
#>  [45] "hash_count"              "hash_min_max"           
#>  [47] "hash_sum"                "index"                  
#>  [49] "index_in"                "index_in_meta_binary"   
#>  [51] "invert"                  "less"                   
#>  [53] "less_equal"              "list_flatten"           
#>  [55] "list_parent_indices"     "list_value_length"      
#>  [57] "match_like"              "match_substring"        
#>  [59] "match_substring_regex"   "mean"                   
#>  [61] "min_max"                 "mode"                   
#>  [63] "multiply"                "multiply_checked"       
#>  [65] "negate"                  "negate_checked"         
#>  [67] "not_equal"               "or"                     
#>  [69] "or_kleene"               "partition_nth_indices"  
#>  [71] "power"                   "power_checked"          
#>  [73] "project"                 "quantile"               
#>  [75] "replace_substring"       "replace_substring_regex"
#>  [77] "sort_indices"            "split_pattern"          
#>  [79] "split_pattern_regex"     "stddev"                 
#>  [81] "string_is_ascii"         "strptime"               
#>  [83] "subtract"                "subtract_checked"       
#>  [85] "sum"                     "take"                   
#>  [87] "tdigest"                 "unique"                 
#>  [89] "utf8_is_alnum"           "utf8_is_alpha"          
#>  [91] "utf8_is_decimal"         "utf8_is_digit"          
#>  [93] "utf8_is_lower"           "utf8_is_numeric"        
#>  [95] "utf8_is_printable"       "utf8_is_space"          
#>  [97] "utf8_is_title"           "utf8_is_upper"          
#>  [99] "utf8_length"             "utf8_lower"             
#> [101] "utf8_ltrim"              "utf8_ltrim_whitespace"  
#> [103] "utf8_reverse"            "utf8_rtrim"             
#> [105] "utf8_rtrim_whitespace"   "utf8_split_whitespace"  
#> [107] "utf8_trim"               "utf8_trim_whitespace"   
#> [109] "utf8_upper"              "value_counts"           
#> [111] "variance"                "xor"